+++ db/query/report.sql
@@ -17,5 +17,5 @@ UPDATE report_forms
 SET
   status = @status
 WHERE
-  id = @id::bigint
+  id = @id::int
 RETURNING *;
